Added socket.io and refactored the server file
