+++ package.json
@@ -14,7 +14,8 @@
     "mongoose": "^5.8.2",
     "mongoose-unique-validator": "^2.0.3",
     "morgan": "^1.9.1",
-    "nodemon": "^2.0.1"
+    "nodemon": "^2.0.1",
+    "socket.io": "^2.3.0"
   },
   "devDependencies": {},
   "scripts": {
