+++ server.js
@@ -1,45 +1,24 @@
-const express = require('express')
-const mongoose = require('mongoose')
-const morgan = require('morgan')
-const bodyParser = require('body-parser')
-const app = express()
+const app = require('./app')
 const config = require('./config')
-const middleware = require('./utils/middleware')
+const http = require('http')
+const server = http.createServer(app)
+const io = require('socket.io')(server)
 
-const userRouter = require('./routes/user')
-const loginRouter = require('./routes/login')
-const quizRouter = require('./routes/quiz')
-const chatRouter = require('./routes/chat')
+app.set('socketio', io)
 
-mongoose.set('useCreateIndex', true)
-
-mongoose
-  .connect(config.MONGODB_URI, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true
-  })
-  .then(() => {
-    console.log('Connected to MongoDB')
-  })
-  .catch(error => {
-    console.log('Error connecting to MongoDB', error.message)
-  })
-
-app.use(bodyParser.json())
-app.use(
-  morgan(':method :url :status :response-time ms - :res[content-length] :body')
-)
-morgan.token('body', (req, res) => JSON.stringify(req.body))
-
-app.use(middleware.tokenExtractor)
-
-app.use('/api/user', userRouter)
-app.use('/api/login', loginRouter)
-app.use('/api/quiz', quizRouter)
-app.use('/api/chat', chatRouter)
-
-app.use(middleware.unknownEndpoint)
-
-app.listen(`${config.PORT}`, () =>
+server.listen(`${config.PORT}`, () =>
   console.log(`Example app listening on port ${config.PORT}!`)
 )
+
+io.on('connection', socket => {
+  console.log('a user connected')
+
+  socket.on('chat', data => {
+    console.log(data)
+    io.emit('chat', data)
+  })
+  socket.on('typing', () => {
+    console.log('User typing')
+    socket.broadcast.emit('typing')
+  })
+})
